[Silver III] Title: 주유소, Time: 20 ms, Memory: 3588 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -13,20 +13,20 @@ int main()
     int n;
     cin >> n;
 
-    vector<int> distance(n - 1);
-    vector<int> cost(n);
-    for (int i = 0; i < n - 1; i++)
+    vector<long long> distance(n - 1);
+    vector<long long> cost(n);
+    for (long long i = 0; i < n - 1; i++)
     {
         cin >> distance[i];
     }
-    for (int i = 0; i < n; i++)
+    for (long long i = 0; i < n; i++)
     {
         cin >> cost[i];
     }
 
-    int temp = cost[0];
-    int answer = 0;
-    for (int i = 1; i < n; i++)
+    long long temp = cost[0];
+    long long answer = 0;
+    for (long long i = 1; i < n; i++)
     {
         answer += temp * distance[i - 1];
         if (cost[i] < temp)
